Move tests to be near the things they are testing
