+++ examples/reproducing/README.md
@@ -1,0 +1,3 @@
+# Reproducible builds
+
+`a` and `b` are the same, and building them both should result in the same bundle
