+++ src/resolver/mod.rs
@@ -1,3 +1,5 @@
+mod test;
+
 use crate::input_options::InputOptions;
 use crate::manifest::{BrowserSubstitution, PackageCache, PackageInfo};
 use crate::path_ext::*;
