+++ src/vlq.rs
@@ -28,3 +28,23 @@ impl Vlq {
         str::from_utf8(&self.buf[0..=l]).unwrap()
     }
 }
+
+mod test {
+    #[test]
+    fn test_vlq() {
+        // 0000000000000000111111111111111122222222222222223333333333333333
+        // 0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef
+        // ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/
+        let mut vlq = super::Vlq::new();
+        assert_eq!(vlq.enc(0), "A");
+        assert_eq!(vlq.enc(1), "C");
+        assert_eq!(vlq.enc(-1), "D");
+        assert_eq!(vlq.enc(5), "K");
+        assert_eq!(vlq.enc(-5), "L");
+        assert_eq!(vlq.enc(15), "e");
+        assert_eq!(vlq.enc(-15), "f");
+        assert_eq!(vlq.enc(16), "gB");
+        assert_eq!(vlq.enc(1876), "o1D"); // 11 10101 0100
+        assert_eq!(vlq.enc(-485223), "v2zd"); // 11101 10011 10110 0111
+    }
+}
